+++ src/main/java/com/octagram/pollet/notification/presentation/controller/NotificationController.java
@@ -1,6 +1,6 @@
-package com.octagram.pollet.notification.presentation;
+package com.octagram.pollet.notification.presentation.controller;
 
-import com.octagram.pollet.notification.application.NotificationService;
+import com.octagram.pollet.notification.application.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
